requirements.txt
youtube_transcript_api
streamlit
google-generativeai
python-dotenv
requests[socks]
flask
gunicorn
importlib-metadata